get history working with login
This means you can bookmark a url, either at a dir in the file
view, or with a particular file open in an app. Logging in via
that url will reopen the file in the app

also make logout clear the url fragment

The key thing here was another vuejs change, watch functions now
take two arguments!