[fix_browser_tests]
This enables the browser tests to be run on new pushed commits via Github actions.
NB: chrome-84 is now _required_ since this is the first version that can both be
run headlessly and download a file without hitting various bugs in the
chromedriver. Currently, only one test is run that signs up a new user and uploads
and downloads a file and checks the contents. The rest of the tests are skipped as
they need to be updated to reflect recent changes in this repo.

More details below:

* fix signup controller and update to python-3
* bump max-users to 100
* Added profile to selenium driver to disable the  file-system app that pops up when downloading a file
* Updated file-upload
* Fixed file-download
* Added smoke test that uploads and downloads a file
* skip broken tests - these need to be updated to reflect the UI changes (mainly changed element ids) since  these were last updated
* update browser-test scripts and  readme
* add github ci for browser tests
* add update docs